+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Track the spawned tasks in progress using `utils::async_work` ([#9546](https://github.com/open-chat-labs/open-chat/pull/9546))
+- Once enabled, register new users in whichever MultiUser canister has the fewest users ([#9579](https://github.com/open-chat-labs/open-chat/pull/9579))
 
 ### Fixed